+++ packages/core/src/components/structural/index.ts
@@ -3,3 +3,5 @@ export { default as BottomAppBar } from './BottomAppBar';
 export { default as NavSidebar } from './NavSidebar';
 export type { NavSidebarProps } from './NavSidebar';
 export { SearchModal } from './SearchModal';
+export { default as AppShellLayout } from './AppShellLayout';
+export type { AppShellLayoutProps } from './AppShellLayout';